requirements.txt
streamlit>=1.28.0
psycopg2-binary>=2.9.0
pandas>=1.5.0
pytz>=2023.3
